requirements.txt
alabaster==0.7.12
anyio==3.3.4
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
argcomplete @ file:///tmp/build/80754af9/argcomplete_1618920853202/work
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
arrow==1.2.1
async-generator==1.10
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
binaryornot==0.4.4
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
blessings==1.7
boolean.py==3.8
boto3==1.19.11
botocore==1.22.11
brotlipy==0.7.0
cached-property==1.5.2
cameo==0.13.5
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1625831764494/work
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
click-configfile==0.2.3
click-default-group==1.2.2
click-log==0.3.2
cobamp==0.2.0
cobra @ file:///home/conda/feedstock_root/build_artifacts/cobra_1626250382024/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
commonmark==0.9.1
configparser==5.1.0
cookiecutter==1.7.3
cryptography @ file:///C:/ci/cryptography_1635366713676/work
cycler==0.11.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
datapackage==1.15.2
debugpy @ file:///C:/ci/debugpy_1629222823515/work
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
depinfo @ file:///home/conda/feedstock_root/build_artifacts/depinfo_1617661968011/work
dill==0.3.4
diskcache @ file:///home/conda/feedstock_root/build_artifacts/diskcache_1611381483884/work
docutils==0.17.1
entrypoints==0.3
Escher==1.7.3
et-xmlfile==1.1.0
future @ file:///D:/bld/future_1635819645981/work
geckopy==2.0.2
gitdb==4.0.9
GitPython==3.1.24
gnomic==1.0.1
goodtables==2.5.4
grako==3.18.1
greenlet==1.1.2
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1609518832927/work
h2 @ file:///D:/bld/h2_1633502854670/work
hpack==4.0.0
httpcore==0.13.7
httpx==0.20.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
ijson==3.1.4
imagesize==1.2.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916842152/work
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1635615662634/work
iniconfig==1.1.1
inspyred==1.0.1
IProgress==0.4
ipykernel @ file:///C:/ci/ipykernel_1633527406033/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1635944342510/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.5
isodate==0.6.0
jedi @ file:///D:/bld/jedi_1635824149087/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1633656206378/work
jinja2-time==0.2.0
jmespath==0.10.0
jmetalpy==1.5.5
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonlines==2.0.0
jsonpointer==2.2
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1630690655192/work
jupyter-core @ file:///C:/ci/jupyter_core_1633402341948/work
jupyter-server @ file:///C:/ci/jupyter_server_1616084265530/work
jupyterlab @ file:///tmp/build/80754af9/jupyterlab_1635799997693/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1633419203660/work
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
linear-tsv==1.1.0
MarkupSafe @ file:///C:/ci/markupsafe_1621528383308/work
matplotlib==3.2.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
memote==0.13.0
mewpy==0.1.9
mistune @ file:///C:/ci/mistune_1594373272338/work
more-itertools==8.10.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1612895720168/work
multiprocess==0.70.12.2
nbclassic @ file:///tmp/build/80754af9/nbclassic_1616085367084/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479160228/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.6.3
notebook @ file:///C:/ci/notebook_1635393701858/work
numexpr==2.7.3
numpy==1.21.4
numpydoc==1.1.0
openpyxl==3.0.9
optlang @ file:///home/conda/feedstock_root/build_artifacts/optlang_1620083108505/work
ordered-set==4.0.2
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
palettable==3.3.0
pandas @ file:///D:/bld/pandas_1634504415850/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1595548966091/work
pathos==0.2.8
patsy==0.5.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
plotly @ file:///Users/nicolas/opt/miniconda3/envs/env/conda-bld/plotly_1630459557113/work/packages/python/plotly/dist/plotly-5.3.1.tar.gz
pluggy==0.13.1
pox==0.3.0
poyo==0.5.0
ppft==1.6.6.4
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1623189609245/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
py==1.11.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.8.2
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pylru==1.2.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyperclip==1.6.0
pyrsistent @ file:///C:/ci/pyrsistent_1636093257833/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytest==4.6.11
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv @ https://github.com/theskumar/python-dotenv/archive/master.zip
python-libsbml @ file:///D:/bld/python-libsbml_1610353695965/work
python-slugify==5.0.2
pytz==2021.3
pywin32==228
pywinpty==0.5.7
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1628276101517/work
reframed==1.2.1
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
rfc3986 @ file:///home/conda/feedstock_root/build_artifacts/rfc3986_1620442452971/work
rich @ file:///D:/bld/rich_1633550169030/work
ruamel.yaml @ file:///D:/bld/ruamel.yaml_1635932740169/work
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scipy==1.7.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
simpleeval==0.9.11
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
sniffio @ file:///C:/ci/sniffio_1614030522573/work
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.26
statistics==1.0.3.5
statsmodels==0.13.0
swiglpk==5.0.3
sympy @ file:///D:/bld/sympy_1633750479590/work
tableschema==1.20.2
tabulator==1.53.5
tenacity @ file:///C:/ci/tenacity_1632759497842/work
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
text-unidecode==1.3
toml==0.10.2
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm==4.62.3
traitlets @ file:///tmp/build/80754af9/traitlets_1632522747050/work
travis-encrypt==1.1.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
unicodecsv==0.14.1
urllib3==1.26.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension==3.5.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
xlrd==2.0.1
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
